+++ internal/engine/eval/trusty/trusty.go
@@ -97,7 +97,7 @@ func (e *Evaluator) Eval(ctx context.Context, pol map[string]any, res *engif.Res
 	}
 
 	logger := zerolog.Ctx(ctx).With().
-		Int32("pull-number", prdeps.Pr.Number).
+		Int64("pull-number", prdeps.Pr.Number).
 		Str("repo-owner", prdeps.Pr.RepoOwner).
 		Str("repo-name", prdeps.Pr.RepoName).
 		Logger()
